requirements-phase2.txt
tldextract==3.6.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
PyYAML==6.0.2
